% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{build_url}
\alias{build_url}
\title{Build a valid URL for grabbing a file from CHCD}
\usage{
build_url(station_id, timecode, year, month = NA)
}
\arguments{
\item{station_id}{The ID of the station to be downloaded}

\item{timecode}{The numeric timecode to be downloaded}

\item{year}{The year of data to be loaded}

\item{month}{The month of the year to download data. Setting to NA
gets all months.}
}
\value{
A URL string
}
\description{
Build a valid URL for grabbing a file from CHCD
}
