% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chcd-package.R
\docType{package}
\name{chcd-package}
\alias{chcd}
\alias{chcd-package}
\title{chcd: Access Canadian Historical Climate Data}
\description{
Provides easy access to historical climate data in Canada from R. Search for weather stations and download raw hourly, daily or monthly weather data across Canada from 1840 to present. Implements public API access as detailed at \url{https://climate.weather.gc.ca}.
}
\author{
\strong{Maintainer}: Dan Prisk \email{dan@prisk.ca} (\href{https://orcid.org/0000-0002-3377-6978}{ORCID}) [copyright holder]

}
\keyword{internal}
