% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{dl_csv}
\alias{dl_csv}
\title{Downloads a csv from the internet}
\usage{
dl_csv(url)
}
\arguments{
\item{url}{A string containing the URL of the CSV to download}
}
\value{
The content of the downloaded CSV
}
\description{
Downloads a csv from the internet
}
