% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{format_columns}
\alias{format_columns}
\title{Directs column formatting requests to the appropriate function.}
\usage{
format_columns(tbl, timecode)
}
\arguments{
\item{tbl}{A tibble containing a single downloaded weather data
file}

\item{timecode}{A timecode representing the form of file downloaded
as generated by get_timecode()}
}
\value{
A single tibble with variable formats standardized
}
\description{
Directs column formatting requests to the appropriate function.
}
