% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{format_columns_daily}
\alias{format_columns_daily}
\title{Sets variables within a single daily weather file to a standardized
format so that files can be combined.}
\usage{
format_columns_daily(tbl)
}
\arguments{
\item{tbl}{A tibble containing a single downloaded daily weather data file}
}
\value{
A single tibble with variables set to standards
}
\description{
Sets variables within a single daily weather file to a standardized
format so that files can be combined.
}
