% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{get_timecode}
\alias{get_timecode}
\title{Converts a text based interval into a CHCD timecode}
\usage{
get_timecode(interval)
}
\arguments{
\item{interval}{A text based interval. Must be one of: "h",
"hourly", "d", "daily", or "m", "monthly".}
}
\value{
A numeric code that can be used in a CHCD URL
}
\description{
Converts a text based interval into a CHCD timecode
}
