% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{stations_near}
\alias{stations_near}
\title{Finds climate stations near to a given location.}
\usage{
stations_near(longitude, latitude, distance)
}
\arguments{
\item{longitude}{The longitude of the point}

\item{latitude}{The latitude of the point}

\item{distance}{The distance in KM from the point to pull in stations}
}
\value{
A tibble containing the id, name, and locations for all
valid stations within the radius of distance from point.
}
\description{
Finds climate stations near to a given location.
}
