% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_climate_file}
\alias{test_climate_file}
\title{Test a climate file has the expected columns}
\usage{
test_climate_file(file_path, expected_columns)
}
\arguments{
\item{file_path}{The path to find the file in the local filesystem}

\item{expected_columns}{A list of expected columns}
}
\value{
A boolean vector
}
\description{
Test a climate file has the expected columns
}
\examples{
file_path <- "/path/to/file/file_name.csv"
expected_columns <- list("Name", "Data Column 1")
test_climate_file(file_path, expected_columns)
}
