% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{build_url}
\alias{build_url}
\title{Build a valid URL for grabbing a file from CHCD}
\usage{
build_url(station_id, interval, year, month = NA)
}
\arguments{
\item{station_id}{The ID of the station to be downloaded}

\item{interval}{String representing the time interval required:
hourly, daily or monthly.}

\item{year}{The year of data to be loaded}

\item{month}{The month of the year to download data. Setting to NA
gets all months.}
}
\value{
A URL string
}
\description{
Build a valid URL for grabbing a file from CHCD
}
\examples{
station_id <- 1706
interval <- "daily"
year <- 1983
month <- NA
build_url(station_id, interval, year, month)
}
