% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.R
\name{get_climatedata}
\alias{get_climatedata}
\title{Get historical climate data for the specified time and place}
\usage{
get_climatedata(place, year, interval)
}
\arguments{
\item{place}{The place that data should be downloaded for. This can
be a climate station ID, or a text place name.}

\item{year}{The year, or years, to get data for. Can be either a
single numeric year or a list of years.}

\item{interval}{The interval that data should be returned for. Must
be one of: "h", "hourly", "m", "monthly", or "y",
"yearly". Defaults to monthly.}
}
\value{
A single tibble containing all the requested data
}
\description{
Get historical climate data for the specified time and place
}
